backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
exceptiongroup==1.2.2
fastapi>=0.68.0
graphql-core==3.2.6
h11==0.14.0
idna==3.10
iso8601==2.1.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
strawberry-graphql>=0.108.0
tortoise-orm>=0.19.0
typing_extensions==4.13.0
uvicorn>=0.15.0
aiomysql>=0.0.22
aerich>=0.6.3  # Para migraciones de base de datos
